REGRESSION(r281389): using font-variant-ligatures causes Unicode bidi isolation control characters to render https://bugs.webkit.org/show_bug.cgi?id=235053 <rdar://problem/87425066>
Reviewed by Darin Adler.

LayoutTests/imported/w3c:

* web-platform-tests/css/css-text/white-space/default-ignorable-complex-expected.html: Added.
* web-platform-tests/css/css-text/white-space/default-ignorable-complex.html: Added.
* web-platform-tests/css/css-text/white-space/default-ignorable-expected.html: Added.
* web-platform-tests/css/css-text/white-space/default-ignorable.html: Added.

Source/WebCore:

The spec says (https://drafts.csswg.org/css-text-3/#white-space-processing)

> unsupported Default_ignorable characters must be ignored for text rendering.

Previously, we were handling this by mapping the characters to the zero width space character. r281389 changed
that to be more faithful to the original string (and not do the mapping). One somewhat unexpected result of
this patch was that, now that we're not mapping these code points to the zero width space character, fewer fonts
on the system will support those code points. It became much more common for no font at all to support the code
point in question. r281389 handled the situation where the code points were supported by a font but shouldn't be
rendered, but didn't handle the case where the code points *weren't* supported by any fonts (in the complex text
codepath). This patch updates that path to handle it.

This patch also replaces our previous hardcoded list of Default_ignorable characters with the proper call to ICU
to check the Unicode property. It does this in both the fast text codepath and the simple text codepath, to make
sure they match.

Tests: imported/w3c/web-platform-tests/css/css-text/white-space/default-ignorable-complex.html
       imported/w3c/web-platform-tests/css/css-text/white-space/default-ignorable.html

* platform/graphics/ComplexTextController.cpp:
(WebCore::ComplexTextController::ComplexTextRun::ComplexTextRun):
* platform/graphics/WidthIterator.cpp:
(WebCore::WidthIterator::applyCSSVisibilityRules):


Canonical link: https://commits.webkit.org/246121@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@288107 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/platform/graphics/WidthIterator.cpp
@@ -616,29 +616,9 @@ void WidthIterator::applyCSSVisibilityRules(GlyphBuffer& glyphBuffer, unsigned g
 
         adjustForSyntheticBold(i);
 
-        if ((characterResponsibleForThisGlyph >= nullCharacter && characterResponsibleForThisGlyph < space)
-            || (characterResponsibleForThisGlyph >= deleteCharacter && characterResponsibleForThisGlyph < noBreakSpace)) {
-            deleteGlyph(i);
-            continue;
-        }
-
-        switch (characterResponsibleForThisGlyph) {
-        case softHyphen:
-        case leftToRightMark:
-        case rightToLeftMark:
-        case leftToRightEmbed:
-        case rightToLeftEmbed:
-        case leftToRightOverride:
-        case rightToLeftOverride:
-        case leftToRightIsolate:
-        case rightToLeftIsolate:
-        case zeroWidthNonJoiner:
-        case zeroWidthJoiner:
-        case popDirectionalFormatting:
-        case popDirectionalIsolate:
-        case firstStrongIsolate:
-        case objectReplacementCharacter:
-        case zeroWidthNoBreakSpace:
+        // https://drafts.csswg.org/css-text-3/#white-space-processing
+        // "Unsupported Default_ignorable characters must be ignored for text rendering."
+        if (FontCascade::isCharacterWhoseGlyphsShouldBeDeletedForTextRendering(characterResponsibleForThisGlyph)) {
             deleteGlyph(i);
             continue;
         }
